api/requirements.txt
fastapi 
pandas
scikit-learn
uvicorn[standard]
azure-servicebus